Move BlockSerdeUtil to SPI
