+++ presto-hive/src/test/java/com/facebook/presto/hive/util/TestSerDeUtils.java
@@ -14,10 +14,10 @@
 package com.facebook.presto.hive.util;
 
 import com.facebook.presto.block.BlockEncodingManager;
-import com.facebook.presto.block.BlockSerdeUtil;
 import com.facebook.presto.spi.block.Block;
 import com.facebook.presto.spi.block.BlockBuilder;
 import com.facebook.presto.spi.block.BlockEncodingSerde;
+import com.facebook.presto.spi.block.BlockSerdeUtil;
 import com.facebook.presto.spi.type.ArrayType;
 import com.facebook.presto.spi.type.RowType;
 import com.facebook.presto.type.TypeRegistry;
